move medical condition classes out of api/data
