+++ src/main/java/com/gregtechceu/gtceu/common/recipe/EnvironmentalHazardCondition.java
@@ -1,6 +1,6 @@
 package com.gregtechceu.gtceu.common.recipe;
 
-import com.gregtechceu.gtceu.api.data.medicalcondition.MedicalCondition;
+import com.gregtechceu.gtceu.api.medicalcondition.MedicalCondition;
 import com.gregtechceu.gtceu.api.machine.trait.RecipeLogic;
 import com.gregtechceu.gtceu.api.recipe.GTRecipe;
 import com.gregtechceu.gtceu.api.recipe.condition.RecipeCondition;
